created a separate swiftUI file for launchScreen.
